# gestepm-rest-api/src/main/resources/apis/gestepm-core-api/api-spec.yaml
openapi: 3.0.3
info:
  title: GestEPM Core API
  version: 0.1.0
  description: GestEPM Core API
  contact:
    name: Carlos Devesa Pérez
servers:
  - url: 'http://localhost:8080'
paths:
...
  /v1/users:
    get:
      summary: V1 - List users
      responses:
        '200':
          $ref: '#/components/responses/UserList'
      operationId: list-users-v1
      description: list-users-v1
      parameters:
        - $ref: ../../base-api-spec.yaml#/components/parameters/api-opt-meta
        - $ref: ../../base-api-spec.yaml#/components/parameters/api-opt-links
        - $ref: ../../base-api-spec.yaml#/components/parameters/api-opt-expand
        - $ref: ../../base-api-spec.yaml#/components/parameters/filter-pagination-offset
        - $ref: ../../base-api-spec.yaml#/components/parameters/filter-pagination-limit
      tags:
        - user-v1
    post:
      summary: ''
      operationId: post-api-users
      responses:
        '200':
          description: OK
      tags:
        - user-v1
    parameters: []
  '/v1/users/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
        description: User id.
    get:
      summary: Your GET endpoint
      tags:
        - user-v1
      responses:
        '200':
          description: OK
      operationId: get-api-users-id
    put:
      summary: ''
      operationId: put-api-users-id
      responses:
        '200':
          description: OK
      tags:
        - user-v1
    delete:
      summary: ''
      operationId: delete-api-users-id
      responses:
        '200':
          description: OK
      tags:
        - user-v1
components:
  schemas:
...
    User:
      title: User
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        mail:
          type: string
        name:
          type: string
        surnames:
          type: string
        country:
          $ref: ../gestepm-master-data-api/api-spec.yaml#/components/schemas/Country
        activityCenter:
          $ref: ../gestepm-master-data-api/api-spec.yaml#/components/schemas/ActivityCenter
        role:
          $ref: ../gestepm-app-config-api/api-spec.yaml#/components/schemas/Role
        level:
          $ref: ../gestepm-master-data-api/api-spec.yaml#/components/schemas/Level
      required:
        - id
        - username
        - mail
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
...
    UserList:
      description: Example response
      content:
        application/json:
          schema:
            type: object
            properties:
              metadata:
                $ref: ../../base-api-spec.yaml#/components/schemas/Meta
              data:
                type: array
                items:
                  $ref: '#/components/schemas/User'
...